requirements.txt
# ============================================================
# APEX Trading OS — Python Dependencies
# Compatible with Python 3.10, 3.11, 3.12
# Install: pip install -r requirements.txt
# ============================================================

# --- Web Framework ---
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-multipart>=0.0.9

# --- Database ---
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
redis>=5.0.0
aioredis>=2.0.1

# --- OANDA ---
oandapyV20>=0.7.2

# --- AI / LLM ---
openai>=1.30.0
anthropic>=0.28.0
google-genai>=0.8.0
# google-generativeai is deprecated — use google-genai above
# DeepSeek, Grok, Qwen all use OpenAI-compatible API (no extra package needed)
langchain>=0.2.0
langchain-openai>=0.1.0

# --- Machine Learning ---
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.3.0

# --- Data & Indicators ---
pandas>=2.0.0
numpy>=1.26.0
# pandas-ta has a broken version string on PyPI — install with the flag below:
#   pip install pandas-ta==0.3.14b0
# OR use the 'ta' package which is a clean drop-in replacement:
ta>=0.11.0
scipy>=1.12.0

# --- Backtesting ---
# vectorbt has Python version conflicts — use 'backtesting' instead (simpler, reliable)
backtesting>=0.3.3

# --- News Sentiment ---
transformers>=4.40.0
# torch: install separately with the right CUDA/CPU build for your machine
#   CPU only: pip install torch --index-url https://download.pytorch.org/whl/cpu
#   Or just: pip install torch  (installs CPU version by default)
vaderSentiment>=3.3.2
newsapi-python>=0.2.7

# --- HTTP / Async ---
httpx>=0.27.0
aiohttp>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# --- Task Queue ---
celery>=5.3.0
flower>=2.0.0

# --- Telegram ---
python-telegram-bot>=21.0.0

# --- Auth & Security ---
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# --- Utilities ---
pydantic>=2.6.0
pydantic-settings>=2.2.0
loguru>=0.7.0
pytz>=2024.1
apscheduler>=3.10.0

# --- Testing ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0

# --- Dev Tools ---
black>=24.0.0
isort>=5.13.0
